silence some warnings with new sdk (mostly signed/unsigned comparison)

+++ Classes/ListController/ServiceListController.m
@@ -631,7 +631,7 @@
 {
 	if(!_supportsNowNext) return;
 #if IS_DEBUG()
-	NSParameterAssert([_mainList count] > indexPath.row);
+	NSParameterAssert([_mainList count] > (NSUInteger)indexPath.row);
 #else
 	if(indexPath.row > [_mainList count])
 		return;
@@ -657,7 +657,7 @@
 		NSObject<EventProtocol> *evt = nil;
 		if(tableView.lastSwipe & swipeTypeRight)
 			evt = (NSObject<EventProtocol > *)[_mainList objectAtIndex: indexPath.row];
-		else if([_subList count] > indexPath.row) // check if we have "next" event, if not the validity check will fail (so ignore the else case)
+		else if([_subList count] > (NSUInteger)indexPath.row) // check if we have "next" event, if not the validity check will fail (so ignore the else case)
 			evt = (NSObject<EventProtocol > *)[_subList objectAtIndex: indexPath.row];
 
 		// FIXME: for convenience reasons a valid service marks an event valid, also if it may
